Updated to 2.0.0
Finally found compatibility with the current release of Valheim.  Cleaned up the modlist with the criteria being that if any mods were preventing compatibility due to lack of updates, or if they were a performance based mod which really didn't help.  Also removed boomstick and odinondemand which weren't utilized very well.